task_management/notification-service/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiokafka==0.12.0
    # via notification-service (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via starlette
asgiref==3.11.0
    # via opentelemetry-instrumentation-asgi
async-timeout==5.0.1
    # via aiokafka
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   loguru
fastapi==0.123.9
    # via notification-service (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.76.0
    # via opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
loguru==0.7.3
    # via notification-service (pyproject.toml)
opentelemetry-api==1.39.0
    # via
    #   notification-service (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp==1.39.0
    # via notification-service (pyproject.toml)
opentelemetry-exporter-otlp-proto-common==1.39.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.39.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.39.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.60b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.60b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.60b0
    # via notification-service (pyproject.toml)
opentelemetry-proto==1.39.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.39.0
    # via
    #   notification-service (pyproject.toml)
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.60b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.60b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==25.0
    # via
    #   aiokafka
    #   opentelemetry-instrumentation
prometheus-client==0.23.1
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.1.0
    # via notification-service (pyproject.toml)
protobuf==6.33.1
    # via
    #   googleapis-common-protos
    #   opentelemetry-proto
pydantic==2.12.5
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via notification-service (pyproject.toml)
python-dotenv==1.2.1
    # via pydantic-settings
requests==2.32.5
    # via opentelemetry-exporter-otlp-proto-http
starlette==0.50.0
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
typing-extensions==4.15.0
    # via
    #   aiokafka
    #   anyio
    #   fastapi
    #   grpcio
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.0
    # via requests
uvicorn==0.38.0
    # via notification-service (pyproject.toml)
win32-setctime==1.2.0
    # via loguru
wrapt==1.17.3
    # via opentelemetry-instrumentation
zipp==3.23.0
    # via importlib-metadata
